+++ kinto_attachment/utils.py
@@ -87,8 +87,10 @@ def patch_record(record, request):
     # Simulate update of fields.
     request.validated = record
     request.body = json.dumps(record).encode('utf-8')
+    request.bound_data['attachment_auto'] = True
     resource = Record(request, context)
     request.current_resource_name = 'record'
+
     try:
         saved = resource.patch()
     except httpexceptions.HTTPNotFound: